+++ packages/excalidraw/actions/actionProperties.tsx
@@ -22,6 +22,7 @@ import {
   arrayToMap,
   getFontFamilyString,
   getLineHeight,
+  isArabicFont,
   isTransparent,
   getStrokeWidthByKey,
   reduceToCommonValue,
@@ -1167,6 +1168,11 @@ export const actionChangeFontFamily = register<{
       appState: {
         ...appState,
         ...nextAppState,
+        ...(nextFontFamily
+          ? isArabicFont(nextFontFamily)
+            ? { lastArabicFontFamily: nextFontFamily }
+            : { lastLatinFontFamily: nextFontFamily }
+          : {}),
       },
       captureUpdate: nextCaptureUpdateAction,
     };